Experiment with worker
The goal was to avoid freezing GUI, epecially on the mobile. This is an experimental version with a limited functionality (only random images). The architecture is far from perferct.

This version generates images in a worker. The GUI waits for using await. It works on the desktop. But the GUI on the mobile (chrome) is still frozen (e.g. the dreaming progress icon is not spinning smoothly or scrolling is impossible).